Add GraphQL loading wrapper for Celo governance page
